Add files via upload
Blue Prism Studio process completed to better handle order processing in the supply chain atmosphere. This process takes an excel spreadsheet of orders and information and with the press of a button all of that information is directed to the corresponding location on the company's processing system. From there, the process will proceed to give you all of the confirmation numbers once completed.